requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
sqlalchemy==1.4.23
pydantic==1.8.2
python-dotenv==0.19.0
websockets==10.0
aiohttp==3.8.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
joblib==1.0.1
pytest==6.2.5
httpx==0.19.0
email-validator==1.1.3 